Comments Addressed
