+++ v0/destinations/engage/utils.js
@@ -1,7 +1,8 @@
 const { set } = require("lodash");
 const {
   getDestinationExternalID,
-  getFieldValueFromMessage
+  getFieldValueFromMessage,
+  flattenMultilevelPayload
 } = require("../../util");
 
 /**
@@ -60,7 +61,7 @@ const getUID = message => {
  * @param {*} message input message
  * @returns Page Event name with category
  */
-const getEvent = message => {
+const generatePageName = message => {
   const { name, category } = message;
   const pageCat = category ? "".concat(category, " ") : "";
   const pageName = name ? "".concat(name, " ") : "";
@@ -72,34 +73,12 @@ const getEvent = message => {
 };
 
 /**
- * Flatens the input payload
- * @param {*} payload Input payload that needs to be flattened
- * @returns the flattened payload at all levels
- */
-const flattenPayload = payload => {
-  const flattenedPayload = {};
-  if (payload) {
-    Object.keys(payload).forEach(v => {
-      if (typeof payload[v] === "object" && !Array.isArray(payload[v])) {
-        const temp = flattenPayload(payload[v]);
-        Object.keys(temp).forEach(i => {
-          flattenedPayload[i] = temp[i];
-        });
-      } else {
-        flattenedPayload[v] = payload[v];
-      }
-    });
-  }
-  return flattenedPayload;
-};
-
-/**
  * @param {*} attributes payload that needs modification and transistion to payload
  * @param {*} specificGenericFields fields that should be overlooked when adding fields to payload
  * @returns payload
  */
 const refinePayload = (attributes, specificGenericFields) => {
-  const flattenedPayload = flattenPayload(attributes);
+  const flattenedPayload = flattenMultilevelPayload(attributes);
   const payload = {};
   Object.keys(flattenedPayload).forEach(v => {
     if (!specificGenericFields.includes(v)) {
@@ -108,4 +87,4 @@ const refinePayload = (attributes, specificGenericFields) => {
   });
   return payload;
 };
-module.exports = { refinePayload, getUID, getEvent, getLists };
+module.exports = { refinePayload, getUID, generatePageName, getLists };
